sih backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
web3==6.12.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pymongo==4.6.0
python-dotenv==1.0.0
